Update SFML to 3.0.2 ver.

+++ include/settings.h
@@ -1,4 +1,4 @@
 #pragma once
 
-constexpr float WINDOW_HEIGHT = 480.f;
-constexpr float WINDOW_WIDTH = 640.f;
+constexpr unsigned WINDOW_HEIGHT = 480.f;
+constexpr unsigned WINDOW_WIDTH = 640.f;
